LayoutTests/fast/url/safari-extension-expected.txt

Verify the canonicalization of Safari Extension URLs.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS src is expected
PASS src is expected
PASS src is expected
FAIL src should be safari-extension://com.f%C3%AAte/. Was safari-extension://com.xn--fte-fma/.
PASS successfullyParsed is true

TEST COMPLETE
